requirements.txt
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-pinecone>=0.2.0
pinecone-client>=5.0.0
gitpython>=3.1.0
python-dotenv>=1.0.0
